fix: 인테리어 마이페이지 경로 수정
