+++ src/components/layout/mypage/interior/MypageInteriorSiderbar.module.scss
