--- src/components/myInteriorpage/MypageInteriorMain.module.css
@@ -1,155 +1,0 @@
-.mainbody {
-  margin: 0 auto;
-  width: 600px;
-  height: auto;
-  text-align: center;
-  padding: 80px;
-}
-
-.containerbox {
-  margin: 0 auto;
-  width: 600px;
-  height: auto;
-  border: 1px solid grey;
-  border-radius: 20px;
-  text-align: center;
-  background-color: white;
-}
-
-.main {
-  padding: 10px;
-  font-weight: bold;
-  color: #6d885d;
-  font-family: Santokki;
-}
-
-.title {
-  padding-top: 20px;
-  text-align: center;
-  font-weight: bold;
-  color: #6d885d;
-}
-
-.td {
-  width: 60px;
-}
-
-.tr {
-  height: 50px;
-  padding: 30px;
-  border-collapse: separate;
-  border-spacing: 10px 10px;
-}
-
-.leftcol {
-  width: 25%;
-  text-align: center;
-  font-weight: bold;
-  color: #6d885d;
-}
-
-.rightcol {
-  width: 65%;
-  text-align: left;
-  padding: 10px;
-}
-
-.upload-button {
-  width: 200px;
-  height: 68px;
-  padding: 10px 20px;
-  background-color: #6d885d;
-  color: white;
-  border: none;
-  border-radius: 17px;
-  font-size: 16px;
-  cursor: pointer;
-}
-
-/* PersonProfilePage.module.css */
-.profilePage {
-  width: 1280px;
-  margin: 0 auto;
-  gap: 40px;
-  padding: 40px 0;
-  background-color: #fffaf4;
-  padding-bottom: 150px;
-  /* 임시용 */
-}
-
-.userInfoCard {
-  width: 283px;
-  height: auto;
-  background-color: #fffaf4;
-  border: 1px solid #6d885d;
-  border-radius: 15px;
-  padding: 20px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  text-align: center;
-}
-
-.profileImage {
-  width: 117px;
-  height: 117px;
-  border-radius: 50%;
-  background-color: #6d885d;
-  margin: 0 auto 15px;
-}
-
-.username {
-  font-size: 20px;
-  font-weight: bold;
-  color: #333;
-  line-height: 1.5;
-}
-
-.userId,
-.userEmail {
-  font-size: 16px;
-  color: #333;
-  line-height: 1.6;
-}
-
-.separator {
-  margin: 20px 0;
-  border: none;
-  border-top: 1px solid #6d885d;
-}
-
-.sectionTitle {
-  font-size: 18px;
-  font-weight: bold;
-  color: #6d885d;
-  font-family: 'Noto Sans KR', sans-serif;
-  padding: 10px;
-}
-
-.sectioncontent {
-  font-size: 16px;
-  color: grey;
-  font-family: 'Noto Sans KR', sans-serif;
-  padding: 10px;
-}
-
-.houseListContainer {
-  flex-grow: 1;
-}
-
-.houseListHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.houseListTitle {
-  font-size: 24px;
-  font-weight: bold;
-  color: #6d885d;
-}
-
-.houseListViewAll {
-  font-size: 16px;
-  color: #6d885d;
-  text-decoration: none;
-}
