--- src/components/mypage/interior/interior/MypagePersonReview.module.scss
@@ -1,70 +1,0 @@
-@import 'variables.scss';
-
-.customTable {
-  width: 100%;
-  text-align: center;
-  border-collapse: separate;
-  table-layout: fixed;
-  border-spacing: 0 16px;
-  margin-bottom: 60px;
-
-  thead {
-    font-weight: bold;
-    border-collapse: collapse;
-  }
-
-  tbody {
-    border-spacing: 0 16px;
-
-    .rowWrap {
-      @include border;
-      display: table-row;
-      box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.25);
-      overflow: hidden;
-      transition: 0.3s ease-in-out;
-
-      &:hover {
-        background-color: #dbe0cb;
-        transform: scale(1);
-      }
-
-      td {
-        border-top: 1px solid $main-color;
-        border-bottom: 1px solid $main-color;
-        padding: 10px;
-        vertical-align: middle;
-        text-align: center;
-
-        &:first-child {
-          border-left: 1px solid $main-color;
-          border-top-left-radius: 10px;
-          border-bottom-left-radius: 10px;
-
-          img {
-            width: 100%;
-            aspect-ratio: 1/1;
-            object-fit: cover;
-          }
-        }
-        &:nth-child(2) {
-          text-align: left;
-        }
-        &:last-of-type {
-          border-right: 1px solid $main-color;
-          border-top-right-radius: 10px;
-          border-bottom-right-radius: 10px;
-        }
-      }
-    }
-  }
-}
-
-.editBtn {
-  @include border;
-  color: $main-color;
-  width: 78px;
-  background-color: white;
-  padding: 10px 0;
-  font-size: 16px;
-  font-weight: bold;
-}
