+++ src/components/mypage/interior/interior/case/MypageInteriorCase.js
