+++ src/components/mypage/interior/interior/case/MypageInteriorCase.module.scss
